Schedule fetches as partition processing completes.
The way we were using Promise.map(...) in BaseConsumer._fetch can
unnecessarily restrict throughput when handlers are relatively slow,
since we need to process *all* partitions before we can start fetching
more partition data.

The intent of this change is to schedule requests to fetch additional
data as soon as we can, using an async.queue to throttle processing
parallelism in place of Promise.map(...)'s `concurrency` option.